[T1-401] refactor: useOutboundTable로 공통 코드 정리
