+++ packages/wms/src/features/outbound/model/index.ts
@@ -1,5 +1,6 @@
 export type * from "./types";
 export * from "./useSearchOutboundPlanForm";
 export * from "./queryHooks";
-export * from "./useOutboundPlanTable";
 export * from "./useCreateOutboundPlanForm";
+export * from "./useSearchOutboundForm";
+export * from "./useOutboundTable";
